% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVB_converting.R
\name{GVB_converting}
\alias{GVB_converting}
\title{GVB_converting() Function}
\usage{
GVB_converting(vep_tab)
}
\arguments{
\item{vep_tab}{a character vector for a file path of the VEP tab output.}
}
\description{
This function converts a VEP tab-delimited output to a GVB input.
}
\examples{
GVB_converting(vep_tab = "VEP-tab_Annotation.txt")
}
