% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVB_histogram.R
\name{GVB_histogram}
\alias{GVB_histogram}
\title{GVB_histogram() Function}
\usage{
GVB_histogram(data, geneSymbol, outputDirectory = NULL)
}
\arguments{
\item{data}{a matrix, the output of GVB_merging() function.}

\item{geneSymbol}{a character vector as gene symbol which is a column name of the 'data'.}

\item{outputDirectory}{a character vector as directory for saving the plot. The file name of plot is geneSymbol_histogram.jpg.}
}
\description{
This function plots the GVB score for one specific gene as a histogram from merged GVB data, which is the output of the GVB() function.
}
\examples{
GVB_histogram(data = merged_gvb, geneSymbol = "Gene1")
}
