% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVB_merging.R
\name{GVB_merging}
\alias{GVB_merging}
\title{GVB_merging() Function}
\usage{
GVB_merging(x)
}
\arguments{
\item{x}{a vector that includes object names of GVB data frames as global variables.}
}
\description{
This function merges GVB data frames of multiple samples derived from GVB() function.
}
\examples{
GVB_merging(x = GVB_DataFrames_Vector)
}
