% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVB_preCalculation.R
\name{GVB_preCalculation}
\alias{GVB_preCalculation}
\title{GVB_preCalculation() Function}
\usage{
GVB_preCalculation(
  input,
  score,
  threshold,
  direction,
  homozygote,
  method,
  indelScore
)
}
\arguments{
\item{input}{a list, the output of GVB_preProcessing() function.}

\item{score}{a character vector for in-silico variant score (e.g.: "SIFT", "PP2", "CADD") which is a column name of the data frame as elements of 'input' (Default = "SIFT").}

\item{threshold}{a numeric vector specifying cut-off of the in-silico variant score. The default value for "SIFT" is 0.7, for "PP2" it's 0.85, and for "CADD" it's 15.}

\item{direction}{The direction of scores to be included based on the threshold. Only one of "or_less" or "or_more" is allowed. "or_less" means <= the threshold and "or_more" means >= the threshold (Default = "or_less").}

\item{homozygote}{logical value. If TRUE, consider homozygotes for GVB calculation. (Default = FALSE).}

\item{method}{allowed values are one of "gm" or "hm". "gm" means geometric mean and "hm" means harmonic mean (Default = "gm").}

\item{indelScore}{a numeric value to assign the in-silico variant score of indel variants when the 'score' parameter is "SIFT" or "PP2" (Default = NULL).}
}
\description{
This function is for calculating the GVB score.
}
\examples{
VDF_list <- GVB_preProcessing(VariantDataFrame)
GVB_preCalculation(input = VDF_list, score = "SIFT", threshold = 0.7, direction = "or_less", homozygote = TRUE, method = "hm", indelScore)
}
