% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVB_preProcessing.R
\name{GVB_preProcessing}
\alias{GVB_preProcessing}
\title{GVB_preProcessing() Function}
\usage{
GVB_preProcessing(x, indel = FALSE)
}
\arguments{
\item{x}{a data frame derived from individual variant annotated file.}

\item{indel}{a logical value. If TRUE, include indel variants (Default = FALSE).}
}
\description{
This function preprocesses the input and converts it to a list where each element is a data frame for each gene before calculating the GVB score.
}
\examples{
GVB_preProcessing(x, indel=FALSE)
}
